server/requirements-mock.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0

# Audio Synthesis
edge-tts>=6.1.10
pydub>=0.25.1

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
